better error handling for state and action scripts

+++ src/system/engines/actionScripts/actionEngine.js
@@ -44,6 +44,19 @@ const transformActionScriptToString = (conditionString) => {
 
 
 const addActionScript = (db,  actionScriptName, topic, script, toTopic ) => {
+  if (typeof(actionScriptName) !== typeof('')){
+    throw (new Error('engines:actionScript:addActionScript actionScriptName must be a string'));
+  }
+  if (typeof(topic) !== typeof('')){
+    throw (new Error('engines:actionScript:addActionScript topic must be a string'));
+  }
+  if (typeof(script) !== typeof('')){
+    throw (new Error('engines:actionScript:addActionScript script must be a string'));
+  }
+  if (typeof(toTopic) !== typeof('')){
+    throw (new Error('engines:actionScript:addActionScript toTopic must be a string'));
+  }
+
   const evalFunction = () => transformActionScriptToString(script)({  });
   actionScripts[actionScriptName] = {
     name: actionScriptName,
@@ -55,6 +68,10 @@ const addActionScript = (db,  actionScriptName, topic, script, toTopic ) => {
 };
 
 const deleteActionScript = (db, actionScriptName) => new Promise((resolve, reject) => {
+  if (typeof(actionScriptName) !== typeof('')){
+    throw (new Error('engines:actionScript:deleteActionScript actionScriptName must be a string'));
+  }
+
   db.open().then(database => {
     database.all(`DELETE FROM action_engine WHERE name = ('${actionScriptName}')`, (err) => {
       delete actionScripts[actionScriptName];
@@ -80,7 +97,7 @@ const getActionScripts = () => actionScripts;
 
 const onMqttTopic = (topic, message) => {
   return Object.keys(actionScripts)
-    .filter(actionScript => path.join('/actions',actionScripts[actionScript].topic) === topic)
+    .filter(actionScript => actionScripts[actionScript].topic === topic)
     .map(actionScript => actionScripts[actionScript])
     .map(script => ({
         toTopic: script.toTopic,
